[REFACTORY] Tela do carrinho [REFACTORY] Footer carrinho [REFACTORY] Tela meus pedidos

+++ app/src/main/java/br/com/ipet/model/entities/Produto.java
@@ -4,5 +4,5 @@ public class Produto {
     public String id;
     public String titulo;
     public String url;
-    public int preco;
+    public float preco;
 }
