+++ app/src/main/java/br/com/ipet/model/entities/Servico.java
@@ -4,5 +4,5 @@ public class Servico {
     public String id;
     public String titulo;
     public String url;
-    public int preco;
+    public float preco;
 }
